KAFKA-4259: Address review comments
