+++ clients/src/main/java/org/apache/kafka/common/security/JaasUtils.java
@@ -20,9 +20,13 @@ import javax.security.auth.login.Configuration;
 import javax.security.auth.login.AppConfigurationEntry;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
+import java.util.Map;
 import java.io.IOException;
 
 import org.apache.kafka.common.KafkaException;
+import org.apache.kafka.common.config.SaslConfigs;
+import org.apache.kafka.common.config.types.Password;
+import org.apache.kafka.common.network.LoginType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -38,16 +42,53 @@ public class JaasUtils {
     public static final String ZK_LOGIN_CONTEXT_NAME_KEY = "zookeeper.sasl.clientconfig";
 
     /**
-     * Construct a JAAS configuration object per kafka jaas configuration file
-     * @param loginContextName
-     * @param key
-     * @return JAAS configuration object
+     * Returns a JAAS Configuration object. For loginType SERVER, default Configuration
+     * is returned. For loginType CLIENT, if Jaas configuration property
+     * {@link SaslConfigs#SASL_JAAS_CONFIG} is specified, the configuration object
+     * is created by parsing the property value. Otherwise, the default Configuration
+     * is returned.
      */
-    public static String jaasConfig(String loginContextName, String key) throws IOException {
-        return jaasConfig(Configuration.getConfiguration(), loginContextName, key);
+    public static Configuration jaasConfig(LoginType loginType, Map<String, ?> configs) {
+        Password jaasConfigArgs = (Password) configs.get(SaslConfigs.SASL_JAAS_CONFIG);
+        if (jaasConfigArgs == null || loginType == LoginType.SERVER)
+            return defaultJaasConfig(loginType);
+        else
+            return new JaasConfig(loginType, jaasConfigArgs.value());
     }
 
-    public static String jaasConfig(Configuration jaasConfig, String loginContextName, String key) throws IOException {
+    private static Configuration defaultJaasConfig(LoginType loginType) {
+        String jaasConfigFile = System.getProperty(JaasUtils.JAVA_LOGIN_CONFIG_PARAM);
+        if (jaasConfigFile == null) {
+            LOG.debug("System property '" + JaasUtils.JAVA_LOGIN_CONFIG_PARAM + "' and Kafka SASL property '" +
+                      SaslConfigs.SASL_JAAS_CONFIG + "' are not set, using default JAAS configuration.");
+        }
+
+        Configuration jaasConfig = Configuration.getConfiguration();
+
+        String loginContextName = loginType.contextName();
+        AppConfigurationEntry[] configEntries = jaasConfig.getAppConfigurationEntry(loginContextName);
+        if (configEntries == null) {
+            String errorMessage;
+            errorMessage = "Could not find a '" + loginContextName + "' entry in the JAAS configuration. System property '" +
+                    JaasUtils.JAVA_LOGIN_CONFIG_PARAM + "' is " + (jaasConfigFile == null ? "not set" : jaasConfigFile);
+            throw new IllegalArgumentException(errorMessage);
+        }
+        return jaasConfig;
+    }
+
+    /**
+     * Returns the configuration option for <code>key</code> from the server login context
+     * of the default JAAS configuration.
+     */
+    public static String defaultServerJaasConfigOption(String key) throws IOException {
+        return jaasConfigOption(Configuration.getConfiguration(), LoginType.SERVER.contextName(), key);
+    }
+
+    /**
+     * Returns the configuration option for <code>key</code> from the login context
+     * <code>loginContextName</code> of the specified JAAS configuration.
+     */
+    public static String jaasConfigOption(Configuration jaasConfig, String loginContextName, String key) throws IOException {
         AppConfigurationEntry[] configurationEntries = jaasConfig.getAppConfigurationEntry(loginContextName);
         if (configurationEntries == null) {
             String errorMessage = "Could not find a '" + loginContextName + "' entry in this configuration.";
