+++ clients/src/main/java/org/apache/kafka/common/security/authenticator/LoginManager.java
@@ -28,7 +28,6 @@ import java.util.EnumMap;
 import java.util.Map;
 
 import org.apache.kafka.common.network.LoginType;
-import org.apache.kafka.common.security.auth.JaasConfigProvider;
 import org.apache.kafka.common.security.auth.Login;
 import org.apache.kafka.common.security.kerberos.KerberosLogin;
 
@@ -63,11 +62,10 @@ public class LoginManager {
      *                  (i.e. consumer and producer)
      * @param configs configuration as key/value pairs
      */
-    public static final LoginManager acquireLoginManager(LoginType loginType, boolean hasKerberos, Map<String, ?> configs) throws IOException, LoginException {
+    public static final LoginManager acquireLoginManager(LoginType loginType, boolean hasKerberos, Map<String, ?> configs, Configuration jaasConfig) throws IOException, LoginException {
         synchronized (LoginManager.class) {
             LoginManager loginManager = CACHED_INSTANCES.get(loginType);
             if (loginManager == null) {
-                Configuration jaasConfig = JaasConfigProvider.jaasConfig(loginType, configs);
                 loginManager = new LoginManager(loginType, hasKerberos, configs, jaasConfig);
                 CACHED_INSTANCES.put(loginType, loginManager);
             }
